update version v0.0.3

+++ src/frontend/src/handler/drop_index.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 PieDb Data
+// Copyright 2022 Piedb Data
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,63 +19,50 @@ use piestream_sqlparser::ast::ObjectName;
 
 use super::RwPgResponse;
 use crate::binder::Binder;
-use crate::catalog::root_catalog::SchemaPath;
+use crate::handler::drop_table::check_source;
 use crate::session::OptimizerContext;
 
 pub async fn handle_drop_index(
     context: OptimizerContext,
-    index_name: ObjectName,
-    if_exists: bool,
+    table_name: ObjectName,
 ) -> Result<RwPgResponse> {
     let session = context.session_ctx;
-    let db_name = session.database();
-    let (schema_name, index_name) =
-        Binder::resolve_schema_qualified_index_name(db_name, index_name)?;
-    let search_path = session.config().get_search_path();
-    let user_name = &session.auth_context().user_name;
-    let schema_path = match schema_name.as_deref() {
-        Some(schema_name) => SchemaPath::Name(schema_name),
-        None => SchemaPath::Path(&search_path, user_name),
-    };
+    let (schema_name, index_name) = Binder::resolve_table_name(table_name)?;
 
+    let catalog_reader = session.env().catalog_reader();
+
+    check_source(catalog_reader, session.clone(), &schema_name, &index_name)?;
     let index_id = {
-        let reader = session.env().catalog_reader().read_guard();
-        match reader.get_index_by_name(db_name, schema_path, &index_name) {
-            Ok((index, _)) => {
-                if session.user_id() != index.index_table.owner {
-                    return Err(PermissionDenied("Do not have the privilege".to_string()).into());
+        let reader = catalog_reader.read_guard();
+        let index_result = reader.get_index_by_name(session.database(), &schema_name, &index_name);
+
+        if index_result.is_err() {
+            let table_result =
+                reader.get_table_by_name(session.database(), &schema_name, &index_name);
+            if table_result.is_ok() {
+                let table = table_result.unwrap();
+                // If associated source is `Some`, then it is a actually a materialized source /
+                // table v2.
+                if table.associated_source_id().is_some() {
+                    return Err(RwError::from(ErrorCode::InvalidInputSyntax(
+                        "Use `DROP TABLE` to drop a table.".to_owned(),
+                    )));
                 }
 
-                index.id
+                return Err(RwError::from(ErrorCode::InvalidInputSyntax(
+                    "Use `DROP MATERIALIZED VIEW` to drop a materialized view.".to_owned(),
+                )));
             }
-            Err(err) => {
-                return match reader.get_table_by_name(db_name, schema_path, &index_name) {
-                    Ok((table, _)) => {
-                        // If associated source is `Some`, then it is a actually a materialized
-                        // source / table v2.
-                        if table.associated_source_id().is_some() {
-                            return Err(RwError::from(ErrorCode::InvalidInputSyntax(
-                                "Use `DROP TABLE` to drop a table.".to_owned(),
-                            )));
-                        }
 
-                        Err(RwError::from(ErrorCode::InvalidInputSyntax(
-                            "Use `DROP MATERIALIZED VIEW` to drop a materialized view.".to_owned(),
-                        )))
-                    }
-                    Err(_) => {
-                        if if_exists {
-                            Ok(RwPgResponse::empty_result_with_notice(
-                                StatementType::DROP_INDEX,
-                                format!("index \"{}\" does not exist, skipping", index_name),
-                            ))
-                        } else {
-                            Err(err)
-                        }
-                    }
-                };
-            }
+            return Err(index_result.unwrap_err());
         }
+
+        let index = index_result.unwrap();
+        if session.user_id() != index.index_table.owner {
+            return Err(PermissionDenied("Do not have the privilege".to_string()).into());
+        }
+
+        index.id
     };
 
     let catalog_writer = session.env().catalog_writer();
@@ -88,7 +75,6 @@ pub async fn handle_drop_index(
 mod tests {
     use piestream_common::catalog::{DEFAULT_DATABASE_NAME, DEFAULT_SCHEMA_NAME};
 
-    use crate::catalog::root_catalog::SchemaPath;
     use crate::test_utils::LocalFrontend;
 
     #[tokio::test]
@@ -102,10 +88,13 @@ mod tests {
         frontend.run_sql(sql_drop_index).await.unwrap();
 
         let session = frontend.session_ref();
-        let catalog_reader = session.env().catalog_reader().read_guard();
-        let schema_path = SchemaPath::Name(DEFAULT_SCHEMA_NAME);
+        let catalog_reader = session.env().catalog_reader();
 
-        let table = catalog_reader.get_table_by_name(DEFAULT_DATABASE_NAME, schema_path, "idx");
-        assert!(table.is_err());
+        let table = catalog_reader
+            .read_guard()
+            .get_table_by_name(DEFAULT_DATABASE_NAME, DEFAULT_SCHEMA_NAME, "idx")
+            .ok()
+            .cloned();
+        assert!(table.is_none());
     }
 }
